fix: format, per class responsibilities, simple division of the load
