+++ src/house_node/__main__.py
@@ -10,7 +10,7 @@ from dotenv import load_dotenv
 from src.openadr_node.models import ReportConfiguration
 from src.openadr_node.models.mqtt_config import MQTTConfig
 from src.openadr_node.models.rest_config import RestApiConfig
-from src.openadr_node.node_manager import NodeManager
+from src.openadr_node.node_controller import NodeController
 
 # Configure logging
 logging.basicConfig(level=logging.INFO)
@@ -50,18 +50,21 @@ def main() -> None:
     ),
   ]
 
-  mqtt_config = MQTTConfig(
-    broker=os.getenv('PRIVATE_MQTT_BROKER_URL'),
-    port=int(os.getenv('PRIVATE_MQTT_PORT')),
-    topic_load_profile=os.getenv('PRIVATE_MQTT_TOPIC_LOAD_PROFILE'),
-    topic_consumption=os.getenv('PRIVATE_MQTT_TOPIC_LOAD_CONSUMPTION'),
-    username=os.getenv('PRIVATE_MQTT_USERNAME'),
-    password=os.getenv('PRIVATE_MQTT_PASSWORD'),
-  )
+  mqtt_config = None
+
+  if os.getenv('PRIVATE_MQTT_BROKER_URL', None):
+    mqtt_config = MQTTConfig(
+      broker=os.getenv('PRIVATE_MQTT_BROKER_URL'),
+      port=int(os.getenv('PRIVATE_MQTT_PORT')),
+      topic_load_profile=os.getenv('PRIVATE_MQTT_TOPIC_LOAD_PROFILE'),
+      topic_consumption=os.getenv('PRIVATE_MQTT_TOPIC_LOAD_CONSUMPTION'),
+      username=os.getenv('PRIVATE_MQTT_USERNAME'),
+      password=os.getenv('PRIVATE_MQTT_PASSWORD'),
+    )
 
   rest_api_config = RestApiConfig(port=int(os.getenv('REST_API_PORT', 5000)))
 
-  node_manager = NodeManager(
+  node_manager = NodeController(
     node_id=os.getenv('NODE_ID'),
     ven_name=os.getenv('VEN_NAME'),
     vtn_url=os.getenv('CONNECT_VTN_URL'),
