+++ src/mock_node/addons/gradio_ui/async_gradio_app.py
@@ -190,7 +190,6 @@ class AsyncGradioApp(AdrBaseConfig):
         'signal_payload': row[SLIDER_VALUE],
       }
       intervals.append(interval)
-
     return intervals
 
   def get_current_allowed_consumption(self) -> str:
